Semester 2/CCMP-606/Assignment 2/assignment_2/requirements.txt
web3==5.12.1
py-solc-x==2.0.2